Move ConfigurationInitializer to correct package
