+++ presto-hive/src/main/java/io/prestosql/plugin/hive/ConfigurationInitializer.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.hive.s3;
+package io.prestosql.plugin.hive;
 
 import org.apache.hadoop.conf.Configuration;
