+++ presto-hive/src/main/java/io/prestosql/plugin/hive/s3/HiveS3Module.java
@@ -16,6 +16,7 @@ package io.prestosql.plugin.hive.s3;
 import com.google.inject.Binder;
 import com.google.inject.Scopes;
 import io.airlift.configuration.AbstractConfigurationAwareModule;
+import io.prestosql.plugin.hive.ConfigurationInitializer;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.common.JavaUtils;
